Address PR comments
* Split into checker/signer files
* Naming convention
